refactor(sdk): split OpenH264 reader utilities
