+++ cvat-video-openh264/src/cvat_video_openh264/__init__.py
@@ -4,7 +4,6 @@
 
 """Decode the constrained MP4/H.264 video chunks generated by CVAT."""
 
-from ._reader import DecoderInfo, iter_frames, resolve_decoder
 from .errors import (
     DecoderIntegrityError,
     DecoderProvisioningError,
@@ -15,6 +14,7 @@ from .errors import (
     VideoDecoderError,
     VideoDecoderUnavailableError,
 )
+from .reader import DecoderInfo, iter_frames, resolve_decoder
 
 __all__ = [
     "DecoderInfo",
