--- cvat-video-openh264/src/cvat_video_openh264/_reader.py
@@ -1,607 +1,0 @@
-# Copyright (C) CVAT.ai Corporation
-#
-# SPDX-License-Identifier: MIT
-
-"""Reader for the constrained MP4/H.264 chunks generated by CVAT."""
-
-from __future__ import annotations
-
-import contextlib
-import ctypes
-import ctypes.util
-import itertools
-import os
-import struct
-from collections.abc import Generator, Iterator
-from pathlib import Path
-from typing import BinaryIO
-
-import PIL.Image
-
-from .ctypes_structs import (
-    BufferInfo,
-    DecoderVTable,
-    DecodingParameters,
-    OpenH264Version,
-    VideoProperty,
-)
-from .errors import DecoderUnavailableError, UnsupportedVideoChunkError
-from .models import AvcConfiguration, Box, DecoderInfo, Sample, VideoTrack
-
-_BOX_HEADER_SIZE = 8
-_EXTENDED_BOX_HEADER_SIZE = 16
-_VISUAL_SAMPLE_ENTRY_SIZE = 78
-_MAX_SAMPLE_COUNT = 1_000_000
-_MAX_SAMPLE_SIZE = 64 * 1024 * 1024
-_ANNEX_B_START_CODE = b"\x00\x00\x00\x01"
-# BT.601 limited→full range LUTs. Expand Y [16..235] and Cb/Cr [16..240] into the
-# full [0..255] components Pillow's YCbCr→RGB conversion assumes.
-_LIMITED_RANGE_LUMA_LUT = tuple(
-    max(0, min(255, round((value - 16) * 255 / 219))) for value in range(256)
-)
-_LIMITED_RANGE_CHROMA_LUT = tuple(
-    max(0, min(255, round(128 + (value - 128) * 255 / 224))) for value in range(256)
-)
-
-
-def _read_exact(file: BinaryIO, size: int) -> bytes:
-    data = file.read(size)
-    if len(data) != size:
-        raise UnsupportedVideoChunkError("The MP4 chunk is truncated")
-
-    return data
-
-
-def _read_at(file: BinaryIO, offset: int, size: int) -> bytes:
-    file.seek(offset)
-    return _read_exact(file, size)
-
-
-def _read_u32(data: bytes, offset: int = 0) -> int:
-    return struct.unpack_from(">I", data, offset)[0]
-
-
-def _read_u64(data: bytes, offset: int = 0) -> int:
-    return struct.unpack_from(">Q", data, offset)[0]
-
-
-def _iter_boxes(file: BinaryIO, start: int, end: int) -> Iterator[Box]:
-    offset = start
-
-    while offset < end:
-        if end - offset < _BOX_HEADER_SIZE:
-            raise UnsupportedVideoChunkError("The MP4 chunk contains an incomplete box header")
-
-        header = _read_at(file, offset, _BOX_HEADER_SIZE)
-        box_size = _read_u32(header)
-        box_type = header[4:8]
-        header_size = _BOX_HEADER_SIZE
-
-        if box_size == 1:
-            if end - offset < _EXTENDED_BOX_HEADER_SIZE:
-                raise UnsupportedVideoChunkError(
-                    "The MP4 chunk contains an incomplete extended box header"
-                )
-
-            box_size = _read_u64(_read_at(file, offset + _BOX_HEADER_SIZE, 8))
-            header_size = _EXTENDED_BOX_HEADER_SIZE
-        elif box_size == 0:
-            box_size = end - offset
-
-        if box_size < header_size or box_size > end - offset:
-            raise UnsupportedVideoChunkError(
-                f"The MP4 chunk contains an invalid {box_type.decode('ascii', 'replace')!r} box"
-            )
-
-        yield Box(
-            type=box_type,
-            offset=offset,
-            payload_offset=offset + header_size,
-            end_offset=offset + box_size,
-        )
-        offset += box_size
-
-
-def _find_box(file: BinaryIO, parent: Box, box_type: bytes) -> Box:
-    matches = [
-        box
-        for box in _iter_boxes(file, parent.payload_offset, parent.end_offset)
-        if box.type == box_type
-    ]
-    if len(matches) != 1:
-        raise UnsupportedVideoChunkError(
-            f"Expected exactly one {box_type.decode('ascii', 'replace')!r} MP4 box"
-        )
-
-    return matches[0]
-
-
-def _parse_avcc(data: bytes) -> AvcConfiguration:
-    if len(data) < 7 or data[0] != 1:
-        raise UnsupportedVideoChunkError("The MP4 chunk has an invalid AVC decoder configuration")
-
-    profile = data[1]
-    profile_compatibility = data[2]
-    if profile != 66 or not profile_compatibility & 0x40:
-        raise UnsupportedVideoChunkError(
-            "Only constrained-baseline H.264 video chunks are supported"
-        )
-
-    nal_length_size = (data[4] & 0x03) + 1
-    offset = 6
-
-    def read_parameter_sets(count: int) -> tuple[bytes, ...]:
-        nonlocal offset
-        parameter_sets = []
-
-        for _ in range(count):
-            if len(data) - offset < 2:
-                raise UnsupportedVideoChunkError(
-                    "The MP4 chunk has a truncated AVC decoder configuration"
-                )
-
-            size = struct.unpack_from(">H", data, offset)[0]
-            offset += 2
-            if not size or size > len(data) - offset:
-                raise UnsupportedVideoChunkError("The MP4 chunk has an invalid AVC parameter set")
-
-            parameter_sets.append(data[offset : offset + size])
-            offset += size
-
-        return tuple(parameter_sets)
-
-    sequence_parameter_sets = read_parameter_sets(data[5] & 0x1F)
-    if offset >= len(data):
-        raise UnsupportedVideoChunkError("The MP4 chunk has no AVC picture parameter sets")
-
-    picture_parameter_set_count = data[offset]
-    offset += 1
-    picture_parameter_sets = read_parameter_sets(picture_parameter_set_count)
-    if not sequence_parameter_sets or not picture_parameter_sets:
-        raise UnsupportedVideoChunkError("The MP4 chunk has incomplete AVC parameter sets")
-
-    return AvcConfiguration(
-        nal_length_size=nal_length_size,
-        sequence_parameter_sets=sequence_parameter_sets,
-        picture_parameter_sets=picture_parameter_sets,
-    )
-
-
-def _parse_sample_description(file: BinaryIO, stsd: Box) -> AvcConfiguration:
-    header = _read_at(file, stsd.payload_offset, 8)
-    entry_count = _read_u32(header, 4)
-    if entry_count != 1:
-        raise UnsupportedVideoChunkError("Expected exactly one MP4 sample description")
-
-    entries_start = stsd.payload_offset + len(header)
-    entries = list(_iter_boxes(file, entries_start, stsd.end_offset))
-    if len(entries) != 1 or entries[0].type != b"avc1":
-        raise UnsupportedVideoChunkError("The MP4 video track is not AVC/H.264")
-
-    avc1 = entries[0]
-    children_start = avc1.payload_offset + _VISUAL_SAMPLE_ENTRY_SIZE
-    if children_start > avc1.end_offset:
-        raise UnsupportedVideoChunkError("The MP4 chunk has a truncated AVC sample entry")
-
-    avcc = _find_box(
-        file,
-        Box(
-            type=b"avc1",
-            offset=avc1.offset,
-            payload_offset=children_start,
-            end_offset=avc1.end_offset,
-        ),
-        b"avcC",
-    )
-    return _parse_avcc(_read_at(file, avcc.payload_offset, avcc.end_offset - avcc.payload_offset))
-
-
-def _parse_sample_sizes(file: BinaryIO, stsz: Box) -> tuple[int, ...]:
-    header = _read_at(file, stsz.payload_offset, 12)
-    common_sample_size = _read_u32(header, 4)
-    sample_count = _read_u32(header, 8)
-
-    if not sample_count or sample_count > _MAX_SAMPLE_COUNT:
-        raise UnsupportedVideoChunkError(f"Unsupported MP4 sample count: {sample_count}")
-
-    if common_sample_size:
-        sample_sizes = (common_sample_size,) * sample_count
-    else:
-        table_size = sample_count * 4
-        if stsz.end_offset - (stsz.payload_offset + 12) != table_size:
-            raise UnsupportedVideoChunkError("The MP4 sample-size table is invalid")
-
-        table = _read_at(file, stsz.payload_offset + 12, table_size)
-        sample_sizes = struct.unpack(f">{sample_count}I", table)
-
-    if any(not size or size > _MAX_SAMPLE_SIZE for size in sample_sizes):
-        raise UnsupportedVideoChunkError("The MP4 chunk contains an invalid AVC sample size")
-
-    return tuple(sample_sizes)
-
-
-def _parse_chunk_offsets(file: BinaryIO, box: Box) -> tuple[int, ...]:
-    entry_count = _read_u32(_read_at(file, box.payload_offset, 8), 4)
-    entry_size = 4 if box.type == b"stco" else 8
-    table_size = entry_count * entry_size
-
-    if not entry_count or box.end_offset - (box.payload_offset + 8) != table_size:
-        raise UnsupportedVideoChunkError("The MP4 chunk-offset table is invalid")
-
-    table = _read_at(file, box.payload_offset + 8, table_size)
-    format_character = "I" if entry_size == 4 else "Q"
-    return tuple(struct.unpack(f">{entry_count}{format_character}", table))
-
-
-def _parse_sample_to_chunk(file: BinaryIO, stsc: Box) -> tuple[tuple[int, int, int], ...]:
-    entry_count = _read_u32(_read_at(file, stsc.payload_offset, 8), 4)
-    table_size = entry_count * 12
-
-    if not entry_count or stsc.end_offset - (stsc.payload_offset + 8) != table_size:
-        raise UnsupportedVideoChunkError("The MP4 sample-to-chunk table is invalid")
-
-    table = _read_at(file, stsc.payload_offset + 8, table_size)
-    entries = tuple(struct.iter_unpack(">III", table))
-    if (
-        entries[0][0] != 1
-        or any(
-            samples_per_chunk == 0 or description_index != 1
-            for _, samples_per_chunk, description_index in entries
-        )
-        or any(current[0] >= following[0] for current, following in itertools.pairwise(entries))
-    ):
-        raise UnsupportedVideoChunkError("The MP4 sample-to-chunk mapping is unsupported")
-
-    return entries
-
-
-def _validate_composition_offsets(file: BinaryIO, stbl: Box) -> None:
-    composition_offset_boxes = [
-        box
-        for box in _iter_boxes(file, stbl.payload_offset, stbl.end_offset)
-        if box.type == b"ctts"
-    ]
-    if not composition_offset_boxes:
-        return
-    if len(composition_offset_boxes) != 1:
-        raise UnsupportedVideoChunkError("The MP4 chunk has multiple composition-offset tables")
-
-    ctts = composition_offset_boxes[0]
-    header = _read_at(file, ctts.payload_offset, 8)
-    version = header[0]
-    entry_count = _read_u32(header, 4)
-    table_size = entry_count * 8
-    if version not in (0, 1) or ctts.end_offset - (ctts.payload_offset + 8) != table_size:
-        raise UnsupportedVideoChunkError("The MP4 composition-offset table is invalid")
-
-    table = _read_at(file, ctts.payload_offset + 8, table_size)
-    offset_format = "i" if version == 1 else "I"
-    if any(offset for _, offset in struct.iter_unpack(f">I{offset_format}", table)):
-        raise UnsupportedVideoChunkError(
-            "H.264 chunks with display-order reordering are not supported"
-        )
-
-
-def _build_samples(
-    sample_sizes: tuple[int, ...],
-    chunk_offsets: tuple[int, ...],
-    sample_to_chunk: tuple[tuple[int, int, int], ...],
-    file_size: int,
-) -> tuple[Sample, ...]:
-    samples = []
-    sample_index = 0
-    mapping_index = 0
-
-    for chunk_number, chunk_offset in enumerate(chunk_offsets, start=1):
-        if (
-            mapping_index + 1 < len(sample_to_chunk)
-            and chunk_number >= sample_to_chunk[mapping_index + 1][0]
-        ):
-            mapping_index += 1
-
-        samples_per_chunk = sample_to_chunk[mapping_index][1]
-        sample_offset = chunk_offset
-
-        for _ in range(samples_per_chunk):
-            if sample_index >= len(sample_sizes):
-                raise UnsupportedVideoChunkError("The MP4 sample tables disagree")
-
-            sample_size = sample_sizes[sample_index]
-            if sample_offset > file_size or sample_size > file_size - sample_offset:
-                raise UnsupportedVideoChunkError("An AVC sample points outside the MP4 chunk")
-
-            samples.append(Sample(offset=sample_offset, size=sample_size))
-            sample_offset += sample_size
-            sample_index += 1
-
-    if sample_index != len(sample_sizes):
-        raise UnsupportedVideoChunkError("The MP4 sample tables disagree")
-
-    return tuple(samples)
-
-
-def _parse_video_track(file: BinaryIO, trak: Box, file_size: int) -> VideoTrack | None:
-    mdia = _find_box(file, trak, b"mdia")
-    hdlr = _find_box(file, mdia, b"hdlr")
-    hdlr_header = _read_at(file, hdlr.payload_offset, 12)
-    if hdlr_header[8:12] != b"vide":
-        return None
-
-    minf = _find_box(file, mdia, b"minf")
-    stbl = _find_box(file, minf, b"stbl")
-    stsd = _find_box(file, stbl, b"stsd")
-    stsz = _find_box(file, stbl, b"stsz")
-    stsc = _find_box(file, stbl, b"stsc")
-
-    chunk_offset_boxes = [
-        box
-        for box in _iter_boxes(file, stbl.payload_offset, stbl.end_offset)
-        if box.type in (b"stco", b"co64")
-    ]
-    if len(chunk_offset_boxes) != 1:
-        raise UnsupportedVideoChunkError("Expected one MP4 chunk-offset table")
-
-    _validate_composition_offsets(file, stbl)
-    sample_sizes = _parse_sample_sizes(file, stsz)
-    chunk_offsets = _parse_chunk_offsets(file, chunk_offset_boxes[0])
-    sample_to_chunk = _parse_sample_to_chunk(file, stsc)
-
-    return VideoTrack(
-        avc_configuration=_parse_sample_description(file, stsd),
-        samples=_build_samples(sample_sizes, chunk_offsets, sample_to_chunk, file_size),
-    )
-
-
-def _read_video_track_from_stream(file: BinaryIO, file_size: int) -> VideoTrack:
-    top_level_boxes = list(_iter_boxes(file, 0, file_size))
-    moov_boxes = [box for box in top_level_boxes if box.type == b"moov"]
-    if len(moov_boxes) != 1:
-        raise UnsupportedVideoChunkError("Expected exactly one MP4 movie box")
-
-    tracks = [
-        track
-        for trak in _iter_boxes(file, moov_boxes[0].payload_offset, moov_boxes[0].end_offset)
-        if trak.type == b"trak"
-        if (track := _parse_video_track(file, trak, file_size)) is not None
-    ]
-    if len(tracks) != 1:
-        raise UnsupportedVideoChunkError("Expected exactly one H.264 video track")
-
-    return tracks[0]
-
-
-def _iter_access_units_from_stream(file: BinaryIO, track: VideoTrack) -> Iterator[bytes]:
-    configuration = track.avc_configuration
-    parameter_sets = configuration.sequence_parameter_sets + configuration.picture_parameter_sets
-
-    for sample_index, sample in enumerate(track.samples):
-        payload = _read_at(file, sample.offset, sample.size)
-        offset = 0
-        access_unit = bytearray()
-
-        if sample_index == 0:
-            for parameter_set in parameter_sets:
-                access_unit += _ANNEX_B_START_CODE
-                access_unit += parameter_set
-
-        while offset < len(payload):
-            if len(payload) - offset < configuration.nal_length_size:
-                raise UnsupportedVideoChunkError("An AVC sample has a truncated NAL length")
-
-            nal_size = int.from_bytes(
-                payload[offset : offset + configuration.nal_length_size], "big"
-            )
-            offset += configuration.nal_length_size
-            if not nal_size or nal_size > len(payload) - offset:
-                raise UnsupportedVideoChunkError("An AVC sample has an invalid NAL length")
-
-            access_unit += _ANNEX_B_START_CODE
-            access_unit += payload[offset : offset + nal_size]
-            offset += nal_size
-
-        yield bytes(access_unit)
-
-
-def _load_library(library_path: str) -> ctypes.CDLL:
-    try:
-        library = ctypes.CDLL(library_path)
-        getattr(library, "WelsCreateDecoder")
-        getattr(library, "WelsDestroyDecoder")
-    except (AttributeError, OSError) as exc:
-        raise DecoderUnavailableError(
-            f"Could not load a compatible OpenH264 library from {library_path!r}: {exc}"
-        ) from exc
-
-    return library
-
-
-def _resolve_decoder_and_library(
-    library_path: os.PathLike[str] | str | None,
-) -> tuple[DecoderInfo, ctypes.CDLL]:
-    configured_path = library_path or os.environ.get("CVAT_OPENH264_LIBRARY")
-    resolved_path = (
-        os.fspath(configured_path) if configured_path else ctypes.util.find_library("openh264")
-    )
-    if not resolved_path:
-        raise DecoderUnavailableError(
-            "OpenH264 is required for video chunks. Set CVAT_OPENH264_LIBRARY to a compatible "
-            "shared library or install OpenH264 on the system."
-        )
-
-    library = _load_library(resolved_path)
-    version = None
-    try:
-        get_version = library.WelsGetCodecVersionEx
-    except AttributeError:
-        pass
-    else:
-        get_version.argtypes = [ctypes.POINTER(OpenH264Version)]
-        get_version.restype = None
-        native_version = OpenH264Version()
-        get_version(ctypes.byref(native_version))
-        version = (native_version.major, native_version.minor, native_version.revision)
-
-    return DecoderInfo(library_path=resolved_path, version=version), library
-
-
-def resolve_decoder(*, library_path: os.PathLike[str] | str | None = None) -> DecoderInfo:
-    """Resolve an explicit, configured, or system OpenH264 library without downloading it."""
-
-    info, _ = _resolve_decoder_and_library(library_path)
-    return info
-
-
-def _copy_plane(pointer: int, width: int, height: int, stride: int) -> PIL.Image.Image:
-    if not pointer or width <= 0 or height <= 0 or stride < width:
-        raise UnsupportedVideoChunkError("OpenH264 returned invalid I420 plane metadata")
-
-    return PIL.Image.frombytes(
-        "L",
-        (width, height),
-        ctypes.string_at(pointer, stride * height),
-        "raw",
-        "L",
-        stride,
-        1,
-    )
-
-
-def _i420_to_rgb(planes: ctypes.Array[ctypes.c_void_p], info: BufferInfo) -> PIL.Image.Image:
-    system_buffer = info.user_data.system_buffer
-    if system_buffer.format != 23:
-        raise UnsupportedVideoChunkError(
-            f"OpenH264 returned unsupported pixel format {system_buffer.format}"
-        )
-
-    width = system_buffer.width
-    height = system_buffer.height
-    chroma_width = (width + 1) // 2
-    chroma_height = (height + 1) // 2
-
-    # CVAT's current chunk writers emit limited-range yuv420p. Pillow's YCbCr conversion
-    # expects full-range components, so expand the planes before converting to RGB.
-    y_plane = _copy_plane(planes[0], width, height, system_buffer.stride[0]).point(
-        _LIMITED_RANGE_LUMA_LUT
-    )
-    u_plane = _copy_plane(planes[1], chroma_width, chroma_height, system_buffer.stride[1]).point(
-        _LIMITED_RANGE_CHROMA_LUT
-    )
-    v_plane = _copy_plane(planes[2], chroma_width, chroma_height, system_buffer.stride[1]).point(
-        _LIMITED_RANGE_CHROMA_LUT
-    )
-
-    return PIL.Image.merge(
-        "YCbCr",
-        (
-            y_plane,
-            u_plane.resize((width, height), PIL.Image.Resampling.BILINEAR),
-            v_plane.resize((width, height), PIL.Image.Resampling.BILINEAR),
-        ),
-    ).convert("RGB")
-
-
-# This decoder be replaced in a text PR with a Cython native OpenH264 boundary
-class _OpenH264Decoder:
-    def __init__(self, decoder_info: DecoderInfo, library: ctypes.CDLL | None = None) -> None:
-        self._library = library if library is not None else _load_library(decoder_info.library_path)
-        self._library.WelsCreateDecoder.argtypes = [ctypes.POINTER(ctypes.c_void_p)]
-        self._library.WelsCreateDecoder.restype = ctypes.c_long
-        self._library.WelsDestroyDecoder.argtypes = [ctypes.c_void_p]
-        self._library.WelsDestroyDecoder.restype = None
-
-        self._decoder = ctypes.c_void_p()
-        result = self._library.WelsCreateDecoder(ctypes.byref(self._decoder))
-        if result or not self._decoder:
-            raise DecoderUnavailableError(f"OpenH264 decoder creation failed with code {result}")
-
-        self._vtable = ctypes.cast(
-            self._decoder, ctypes.POINTER(ctypes.POINTER(DecoderVTable))
-        ).contents.contents
-        parameters = DecodingParameters()
-        parameters.video_property.size = ctypes.sizeof(VideoProperty)
-        parameters.video_property.bitstream_type = 0
-
-        result = self._vtable.initialize(self._decoder, ctypes.byref(parameters))
-        if result:
-            self._library.WelsDestroyDecoder(self._decoder)
-            self._decoder = ctypes.c_void_p()
-            raise DecoderUnavailableError(
-                f"OpenH264 decoder initialization failed with code {result}"
-            )
-
-    def decode(self, access_unit: bytes) -> PIL.Image.Image | None:
-        source = (ctypes.c_ubyte * len(access_unit)).from_buffer_copy(access_unit)
-        planes = (ctypes.c_void_p * 3)()
-        buffer_info = BufferInfo()
-        state = self._vtable.decode_frame_no_delay(
-            self._decoder,
-            source,
-            len(source),
-            planes,
-            ctypes.byref(buffer_info),
-        )
-        if state & ~0x01:
-            raise UnsupportedVideoChunkError(
-                f"OpenH264 failed to decode the video chunk (state 0x{state:x})"
-            )
-        if not buffer_info.buffer_status:
-            return None
-
-        return _i420_to_rgb(planes, buffer_info)
-
-    def close(self) -> None:
-        if not self._decoder:
-            return
-
-        try:
-            self._vtable.uninitialize(self._decoder)
-        finally:
-            self._library.WelsDestroyDecoder(self._decoder)
-            self._decoder = ctypes.c_void_p()
-
-
-def iter_frames(
-    path: os.PathLike[str] | str,
-    *,
-    library_path: os.PathLike[str] | str | None = None,
-) -> Generator[PIL.Image.Image, None, None]:
-    """
-    Decode a CVAT-generated constrained-baseline MP4 chunk sequentially.
-
-    The frame-count integrity check runs only when this iterator is exhausted normally.
-    This adapter deliberately has no codec downloader or general-purpose media fallback.
-    """
-
-    chunk_path = Path(path)
-
-    try:
-        file_size = chunk_path.stat().st_size
-        file = chunk_path.open("rb")
-    except OSError as exc:
-        raise UnsupportedVideoChunkError(
-            f"Could not read video chunk {chunk_path!s}: {exc}"
-        ) from exc
-
-    # Parsing and access-unit streaming intentionally share this handle. Closing the
-    # returned generator exits this context and releases both the decoder and the file.
-    with contextlib.closing(file):
-        track = _read_video_track_from_stream(file, file_size)
-        decoder_info, library = _resolve_decoder_and_library(library_path)
-        decoder = _OpenH264Decoder(decoder_info, library=library)
-        decoded_frame_count = 0
-
-        try:
-            access_units = _iter_access_units_from_stream(file, track)
-            with contextlib.closing(access_units):
-                for access_unit in access_units:
-                    image = decoder.decode(access_unit)
-                    if image is not None:
-                        decoded_frame_count += 1
-                        yield image
-        finally:
-            decoder.close()
-
-    if decoded_frame_count != len(track.samples):
-        raise UnsupportedVideoChunkError(
-            f"Decoded {decoded_frame_count} frames from {len(track.samples)} AVC samples"
-        )
