+++ cvat-video-openh264/src/cvat_video_openh264/reader.py
@@ -1,0 +1,79 @@
+# Copyright (C) CVAT.ai Corporation
+#
+# SPDX-License-Identifier: MIT
+
+"""Reader for the constrained MP4/H.264 chunks generated by CVAT."""
+
+from __future__ import annotations
+
+import contextlib
+import os
+from collections.abc import Generator
+from pathlib import Path
+
+import PIL.Image
+
+from .errors import UnsupportedVideoChunkError
+from .models import DecoderInfo
+from .utils import (
+    OpenH264Decoder,
+    iter_access_units_from_stream,
+    read_video_track_from_stream,
+    resolve_decoder_and_library,
+)
+
+__all__ = ["DecoderInfo", "iter_frames", "resolve_decoder"]
+
+
+def resolve_decoder(*, library_path: os.PathLike[str] | str | None = None) -> DecoderInfo:
+    """Resolve an explicit, configured, or system OpenH264 library without downloading it."""
+
+    info, _ = resolve_decoder_and_library(library_path)
+    return info
+
+
+def iter_frames(
+    path: os.PathLike[str] | str,
+    *,
+    library_path: os.PathLike[str] | str | None = None,
+) -> Generator[PIL.Image.Image, None, None]:
+    """
+    Decode a CVAT-generated constrained-baseline MP4 chunk sequentially.
+
+    The frame-count integrity check runs only when this iterator is exhausted normally.
+    This adapter deliberately has no codec downloader or general-purpose media fallback.
+    """
+
+    chunk_path = Path(path)
+
+    try:
+        file_size = chunk_path.stat().st_size
+        file = chunk_path.open("rb")
+    except OSError as exc:
+        raise UnsupportedVideoChunkError(
+            f"Could not read video chunk {chunk_path!s}: {exc}"
+        ) from exc
+
+    # Parsing and access-unit streaming intentionally share this handle. Closing the
+    # returned generator exits this context and releases both the decoder and the file.
+    with contextlib.closing(file):
+        track = read_video_track_from_stream(file, file_size)
+        decoder_info, library = resolve_decoder_and_library(library_path)
+        decoder = OpenH264Decoder(decoder_info, library=library)
+        decoded_frame_count = 0
+
+        try:
+            access_units = iter_access_units_from_stream(file, track)
+            with contextlib.closing(access_units):
+                for access_unit in access_units:
+                    image = decoder.decode(access_unit)
+                    if image is not None:
+                        decoded_frame_count += 1
+                        yield image
+        finally:
+            decoder.close()
+
+    if decoded_frame_count != len(track.samples):
+        raise UnsupportedVideoChunkError(
+            f"Decoded {decoded_frame_count} frames from {len(track.samples)} AVC samples"
+        )
