+++ cvat-video-openh264/src/cvat_video_openh264/utils/mp4.py
@@ -1,0 +1,410 @@
+# Copyright (C) CVAT.ai Corporation
+#
+# SPDX-License-Identifier: MIT
+
+"""Parse the constrained MP4/H.264 chunks generated by CVAT."""
+
+from __future__ import annotations
+
+import itertools
+import struct
+from collections.abc import Iterator
+from typing import BinaryIO
+
+from ..errors import UnsupportedVideoChunkError
+from ..models import AvcConfiguration, Box, Sample, VideoTrack
+
+_BOX_HEADER_SIZE = 8
+_EXTENDED_BOX_HEADER_SIZE = 16
+_VISUAL_SAMPLE_ENTRY_SIZE = 78
+MAX_SAMPLE_COUNT = 1_000_000
+MAX_SAMPLE_SIZE = 256 * 1024 * 1024
+ANNEX_B_START_CODE = b"\x00\x00\x00\x01"
+
+
+def read_exact(file: BinaryIO, size: int) -> bytes:
+    """Read exactly ``size`` bytes or reject the truncated chunk."""
+
+    data = file.read(size)
+    if len(data) != size:
+        raise UnsupportedVideoChunkError("The MP4 chunk is truncated")
+
+    return data
+
+
+def read_at(file: BinaryIO, offset: int, size: int) -> bytes:
+    """Read an exact byte range from a seekable stream."""
+
+    file.seek(offset)
+    return read_exact(file, size)
+
+
+def read_u32(data: bytes, offset: int = 0) -> int:
+    """Read one big-endian unsigned 32-bit integer."""
+
+    return struct.unpack_from(">I", data, offset)[0]
+
+
+def read_u64(data: bytes, offset: int = 0) -> int:
+    """Read one big-endian unsigned 64-bit integer."""
+
+    return struct.unpack_from(">Q", data, offset)[0]
+
+
+def iter_boxes(file: BinaryIO, start: int, end: int) -> Iterator[Box]:
+    """Yield the MP4 boxes contained in the half-open byte range."""
+
+    offset = start
+
+    while offset < end:
+        if end - offset < _BOX_HEADER_SIZE:
+            raise UnsupportedVideoChunkError("The MP4 chunk contains a truncated box header")
+
+        header = read_at(file, offset, _BOX_HEADER_SIZE)
+        box_size = read_u32(header)
+        box_type = header[4:8]
+        header_size = _BOX_HEADER_SIZE
+
+        if box_size == 1:
+            if end - offset < _EXTENDED_BOX_HEADER_SIZE:
+                raise UnsupportedVideoChunkError(
+                    "The MP4 chunk contains a truncated extended box header"
+                )
+
+            box_size = read_u64(read_at(file, offset + _BOX_HEADER_SIZE, 8))
+            header_size = _EXTENDED_BOX_HEADER_SIZE
+        elif box_size == 0:
+            box_size = end - offset
+
+        if box_size < header_size or box_size > end - offset:
+            raise UnsupportedVideoChunkError(
+                f"The MP4 chunk contains an invalid {box_type.decode('ascii', 'replace')!r} box"
+            )
+
+        yield Box(
+            type=box_type,
+            offset=offset,
+            payload_offset=offset + header_size,
+            end_offset=offset + box_size,
+        )
+        offset += box_size
+
+
+def find_box(file: BinaryIO, parent: Box, box_type: bytes) -> Box:
+    """Find the single child of ``parent`` with the requested type."""
+
+    matches = [
+        box
+        for box in iter_boxes(file, parent.payload_offset, parent.end_offset)
+        if box.type == box_type
+    ]
+    if len(matches) != 1:
+        raise UnsupportedVideoChunkError(
+            f"Expected exactly one {box_type.decode('ascii', 'replace')!r} MP4 box"
+        )
+
+    return matches[0]
+
+
+def parse_avcc(data: bytes) -> AvcConfiguration:
+    """Parse and validate an AVC decoder configuration record."""
+
+    if len(data) < 7 or data[0] != 1:
+        raise UnsupportedVideoChunkError("The MP4 chunk has an invalid AVC decoder configuration")
+
+    profile = data[1]
+    profile_compatibility = data[2]
+    if profile != 66 or not profile_compatibility & 0x40:
+        raise UnsupportedVideoChunkError(
+            "Only constrained-baseline H.264 video chunks are supported"
+        )
+
+    nal_length_size = (data[4] & 0x03) + 1
+    offset = 6
+
+    def read_parameter_sets(count: int) -> tuple[bytes, ...]:
+        nonlocal offset
+        parameter_sets = []
+
+        for _ in range(count):
+            if len(data) - offset < 2:
+                raise UnsupportedVideoChunkError(
+                    "The MP4 chunk has a truncated AVC decoder configuration"
+                )
+
+            size = struct.unpack_from(">H", data, offset)[0]
+            offset += 2
+            if not size or size > len(data) - offset:
+                raise UnsupportedVideoChunkError("The MP4 chunk has an invalid AVC parameter set")
+
+            parameter_sets.append(data[offset : offset + size])
+            offset += size
+
+        return tuple(parameter_sets)
+
+    sequence_parameter_sets = read_parameter_sets(data[5] & 0x1F)
+    if offset >= len(data):
+        raise UnsupportedVideoChunkError("The MP4 chunk has no AVC picture parameter sets")
+
+    picture_parameter_set_count = data[offset]
+    offset += 1
+    picture_parameter_sets = read_parameter_sets(picture_parameter_set_count)
+    if not sequence_parameter_sets or not picture_parameter_sets:
+        raise UnsupportedVideoChunkError("The MP4 chunk has incomplete AVC parameter sets")
+
+    return AvcConfiguration(
+        nal_length_size=nal_length_size,
+        sequence_parameter_sets=sequence_parameter_sets,
+        picture_parameter_sets=picture_parameter_sets,
+    )
+
+
+def parse_sample_description(file: BinaryIO, stsd: Box) -> AvcConfiguration:
+    """Read the single supported ``avc1`` sample-description entry."""
+
+    header = read_at(file, stsd.payload_offset, 8)
+    entry_count = read_u32(header, 4)
+    if entry_count != 1:
+        raise UnsupportedVideoChunkError("Expected exactly one MP4 sample description")
+
+    entries_start = stsd.payload_offset + len(header)
+    entries = list(iter_boxes(file, entries_start, stsd.end_offset))
+    if len(entries) != 1 or entries[0].type != b"avc1":
+        raise UnsupportedVideoChunkError("The MP4 video track is not AVC/H.264")
+
+    avc1 = entries[0]
+    children_start = avc1.payload_offset + _VISUAL_SAMPLE_ENTRY_SIZE
+    if children_start > avc1.end_offset:
+        raise UnsupportedVideoChunkError("The MP4 chunk has a truncated AVC sample entry")
+
+    avcc = find_box(
+        file,
+        Box(
+            type=b"avc1",
+            offset=avc1.offset,
+            payload_offset=children_start,
+            end_offset=avc1.end_offset,
+        ),
+        b"avcC",
+    )
+    return parse_avcc(read_at(file, avcc.payload_offset, avcc.end_offset - avcc.payload_offset))
+
+
+def parse_sample_sizes(file: BinaryIO, stsz: Box) -> tuple[int, ...]:
+    """Parse and bound-check the MP4 sample-size table."""
+
+    header = read_at(file, stsz.payload_offset, 12)
+    common_sample_size = read_u32(header, 4)
+    sample_count = read_u32(header, 8)
+
+    if not sample_count or sample_count > MAX_SAMPLE_COUNT:
+        raise UnsupportedVideoChunkError(f"Unsupported MP4 sample count: {sample_count}")
+
+    if common_sample_size:
+        sample_sizes = (common_sample_size,) * sample_count
+    else:
+        table_size = sample_count * 4
+        if stsz.end_offset - (stsz.payload_offset + 12) != table_size:
+            raise UnsupportedVideoChunkError("The MP4 sample-size table is invalid")
+
+        table = read_at(file, stsz.payload_offset + 12, table_size)
+        sample_sizes = struct.unpack(f">{sample_count}I", table)
+
+    if any(not size or size > MAX_SAMPLE_SIZE for size in sample_sizes):
+        raise UnsupportedVideoChunkError("The MP4 chunk contains an invalid AVC sample size")
+
+    return tuple(sample_sizes)
+
+
+def parse_chunk_offsets(file: BinaryIO, box: Box) -> tuple[int, ...]:
+    """Parse an ``stco`` or ``co64`` chunk-offset table."""
+
+    entry_count = read_u32(read_at(file, box.payload_offset, 8), 4)
+    entry_size = 4 if box.type == b"stco" else 8
+    table_size = entry_count * entry_size
+
+    if not entry_count or box.end_offset - (box.payload_offset + 8) != table_size:
+        raise UnsupportedVideoChunkError("The MP4 chunk-offset table is invalid")
+
+    table = read_at(file, box.payload_offset + 8, table_size)
+    format_character = "I" if entry_size == 4 else "Q"
+    return tuple(struct.unpack(f">{entry_count}{format_character}", table))
+
+
+def parse_sample_to_chunk(file: BinaryIO, stsc: Box) -> tuple[tuple[int, int, int], ...]:
+    """Parse and validate the MP4 sample-to-chunk mapping."""
+
+    entry_count = read_u32(read_at(file, stsc.payload_offset, 8), 4)
+    table_size = entry_count * 12
+
+    if not entry_count or stsc.end_offset - (stsc.payload_offset + 8) != table_size:
+        raise UnsupportedVideoChunkError("The MP4 sample-to-chunk table is invalid")
+
+    table = read_at(file, stsc.payload_offset + 8, table_size)
+    entries = tuple(struct.iter_unpack(">III", table))
+    if (
+        entries[0][0] != 1
+        or any(
+            samples_per_chunk == 0 or description_index != 1
+            for _, samples_per_chunk, description_index in entries
+        )
+        or any(current[0] >= following[0] for current, following in itertools.pairwise(entries))
+    ):
+        raise UnsupportedVideoChunkError("The MP4 sample-to-chunk mapping is unsupported")
+
+    return entries
+
+
+def validate_composition_offsets(file: BinaryIO, stbl: Box) -> None:
+    """Reject composition offsets that would require display-order reordering."""
+
+    composition_offset_boxes = [
+        box for box in iter_boxes(file, stbl.payload_offset, stbl.end_offset) if box.type == b"ctts"
+    ]
+    if not composition_offset_boxes:
+        return
+    if len(composition_offset_boxes) != 1:
+        raise UnsupportedVideoChunkError("The MP4 chunk has multiple composition-offset tables")
+
+    ctts = composition_offset_boxes[0]
+    header = read_at(file, ctts.payload_offset, 8)
+    version = header[0]
+    entry_count = read_u32(header, 4)
+    table_size = entry_count * 8
+    if version not in (0, 1) or ctts.end_offset - (ctts.payload_offset + 8) != table_size:
+        raise UnsupportedVideoChunkError("The MP4 composition-offset table is invalid")
+
+    table = read_at(file, ctts.payload_offset + 8, table_size)
+    offset_format = "i" if version == 1 else "I"
+    if any(offset for _, offset in struct.iter_unpack(f">I{offset_format}", table)):
+        raise UnsupportedVideoChunkError(
+            "H.264 chunks with display-order reordering are not supported"
+        )
+
+
+def build_samples(
+    sample_sizes: tuple[int, ...],
+    chunk_offsets: tuple[int, ...],
+    sample_to_chunk: tuple[tuple[int, int, int], ...],
+    file_size: int,
+) -> tuple[Sample, ...]:
+    """Build checked byte ranges from the MP4 sample tables."""
+
+    samples = []
+    sample_index = 0
+    mapping_index = 0
+
+    for chunk_number, chunk_offset in enumerate(chunk_offsets, start=1):
+        if (
+            mapping_index + 1 < len(sample_to_chunk)
+            and chunk_number >= sample_to_chunk[mapping_index + 1][0]
+        ):
+            mapping_index += 1
+
+        samples_per_chunk = sample_to_chunk[mapping_index][1]
+        sample_offset = chunk_offset
+
+        for _ in range(samples_per_chunk):
+            if sample_index >= len(sample_sizes):
+                raise UnsupportedVideoChunkError("The MP4 sample tables disagree")
+
+            sample_size = sample_sizes[sample_index]
+            if sample_offset > file_size or sample_size > file_size - sample_offset:
+                raise UnsupportedVideoChunkError("An AVC sample points outside the MP4 chunk")
+
+            samples.append(Sample(offset=sample_offset, size=sample_size))
+            sample_offset += sample_size
+            sample_index += 1
+
+    if sample_index != len(sample_sizes):
+        raise UnsupportedVideoChunkError("The MP4 sample tables disagree")
+
+    return tuple(samples)
+
+
+def parse_video_track(file: BinaryIO, trak: Box, file_size: int) -> VideoTrack | None:
+    """Parse one MP4 track, returning ``None`` for a non-video track."""
+
+    mdia = find_box(file, trak, b"mdia")
+    hdlr = find_box(file, mdia, b"hdlr")
+    hdlr_header = read_at(file, hdlr.payload_offset, 12)
+    if hdlr_header[8:12] != b"vide":
+        return None
+
+    minf = find_box(file, mdia, b"minf")
+    stbl = find_box(file, minf, b"stbl")
+    stsd = find_box(file, stbl, b"stsd")
+    stsz = find_box(file, stbl, b"stsz")
+    stsc = find_box(file, stbl, b"stsc")
+
+    chunk_offset_boxes = [
+        box
+        for box in iter_boxes(file, stbl.payload_offset, stbl.end_offset)
+        if box.type in (b"stco", b"co64")
+    ]
+    if len(chunk_offset_boxes) != 1:
+        raise UnsupportedVideoChunkError("Expected one MP4 chunk-offset table")
+
+    validate_composition_offsets(file, stbl)
+    sample_sizes = parse_sample_sizes(file, stsz)
+    chunk_offsets = parse_chunk_offsets(file, chunk_offset_boxes[0])
+    sample_to_chunk = parse_sample_to_chunk(file, stsc)
+
+    return VideoTrack(
+        avc_configuration=parse_sample_description(file, stsd),
+        samples=build_samples(sample_sizes, chunk_offsets, sample_to_chunk, file_size),
+    )
+
+
+def read_video_track_from_stream(file: BinaryIO, file_size: int) -> VideoTrack:
+    """Read the single supported H.264 video track from an open chunk stream."""
+
+    top_level_boxes = list(iter_boxes(file, 0, file_size))
+    moov_boxes = [box for box in top_level_boxes if box.type == b"moov"]
+    if len(moov_boxes) != 1:
+        raise UnsupportedVideoChunkError("Expected exactly one MP4 movie box")
+
+    tracks = [
+        track
+        for trak in iter_boxes(file, moov_boxes[0].payload_offset, moov_boxes[0].end_offset)
+        if trak.type == b"trak"
+        if (track := parse_video_track(file, trak, file_size)) is not None
+    ]
+    if len(tracks) != 1:
+        raise UnsupportedVideoChunkError("Expected exactly one H.264 video track")
+
+    return tracks[0]
+
+
+def iter_access_units_from_stream(file: BinaryIO, track: VideoTrack) -> Iterator[bytes]:
+    """Yield the track's length-prefixed AVC samples as Annex-B access units."""
+
+    configuration = track.avc_configuration
+    parameter_sets = configuration.sequence_parameter_sets + configuration.picture_parameter_sets
+
+    for sample_index, sample in enumerate(track.samples):
+        payload = read_at(file, sample.offset, sample.size)
+        offset = 0
+        access_unit = bytearray()
+
+        if sample_index == 0:
+            for parameter_set in parameter_sets:
+                access_unit += ANNEX_B_START_CODE
+                access_unit += parameter_set
+
+        while offset < len(payload):
+            if len(payload) - offset < configuration.nal_length_size:
+                raise UnsupportedVideoChunkError("An AVC sample has a truncated NAL length")
+
+            nal_size = int.from_bytes(
+                payload[offset : offset + configuration.nal_length_size], "big"
+            )
+            offset += configuration.nal_length_size
+            if not nal_size or nal_size > len(payload) - offset:
+                raise UnsupportedVideoChunkError("An AVC sample has an invalid NAL length")
+
+            access_unit += ANNEX_B_START_CODE
+            access_unit += payload[offset : offset + nal_size]
+            offset += nal_size
+
+        yield bytes(access_unit)
